requirements.txt
fire==0.6.0
fonttools==4.51.0
lxml==5.2.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
pdf2docx==0.5.8
pillow==10.3.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
PyPDF4==1.27.0
python-docx==1.1.0
six==1.16.0
termcolor==2.4.0
traits==6.4.3
typing_extensions==4.11.0
wheel==0.43.0
